refactor(config): delete useless code and move makeSidebar from common.ts to serverUtils
