+++ docs/.vitepress/config/common.ts
@@ -1,9 +1,3 @@
-import fs from 'fs'
-import { parse } from 'yaml'
-import YAML from 'js-yaml'
-import { enConfig } from './en'
-import { zhConfig } from '../../zh/config/index'
-
 export const common = async () => {
   return {
     appearance: false,
@@ -21,7 +15,6 @@ export const common = async () => {
       logo: 'logo-text-tinted.png',
       copyright: '©Copyright 2022 Greptime Inc. All Rights Reserved',
       email: 'marketing@greptime.com',
-      // sidebar: await makeSidebar('en'),
       editLink: {
         pattern: 'https://github.com/GreptimeTeam/docs/blob/main/docs/:path',
       },
@@ -35,43 +28,4 @@ export const common = async () => {
     },
     cleanUrls: 'without-subfolders',
   }
-}
-
-export async function makeSidebar(lang) {
-  const langPath = lang !== 'en' ? `/${lang}` : ''
-
-  const summary = YAML.load(fs.readFileSync(`docs/summary.yml`), 'utf8')
-  const summaryI18n = langPath ? YAML.load(fs.readFileSync(`docs${langPath}/summary-i18n.yml`), 'utf8') : null
-
-  function makeSidebarItem(items, path, level = 0) {
-    if (Array.isArray(items)) {
-      return items.map(item => makeSidebarItem(item, path, level + 1))
-    } else if (typeof items === 'object') {
-      let title = Object.keys(items)[0]
-      let content = Object.values(items)[0]
-
-      if (summaryI18n && !summaryI18n[title]) {
-        return {}
-      }
-
-      return {
-        text: summaryI18n?.[title] || title.replace(/-/g, ' '),
-        items: content.map(item => makeSidebarItem(item, `${path}/${title}`, level + 1)),
-        collapsed: level > 1,
-      }
-    } else {
-      try {
-        let link = `${path}/${items}`.toLocaleLowerCase()
-        let file = fs.readFileSync(`docs${langPath}${link}.md`, 'utf-8')
-        return {
-          text: file.split('\n')[0].replace('# ', ''),
-          link,
-        }
-      } catch (error) {
-        return {}
-      }
-    }
-  }
-
-  return makeSidebarItem(summary, '')
 }
