+++ docs/.vitepress/theme/serverUtils.ts
@@ -1,75 +1,41 @@
-import globby from 'globby'
-import matter from 'gray-matter'
-import md5 from 'md5'
 import fs from 'fs-extra'
-import MarkdownIt from 'markdown-it'
+import YAML from 'js-yaml'
 
-const md = new MarkdownIt()
-const { spawn } = require('cross-spawn')
-const dayjs = require('dayjs')
-const gitDateExtractor = require('git-date-extractor')
+export async function makeSidebar(lang) {
+  const langPath = lang !== 'en' ? `/${lang}` : ''
 
-export async function getBlogs() {
-	let paths = await getPostMDFilePaths()
-	let posts = await Promise.all(
-		paths.map(async (item) => {
-			const gitInfo = await getGitTimestamp(item)
-			const file = await fs.readFile(item, 'utf-8')
-			const { data, content } = matter(file)
-			const article = md.render(content)
-			return {
-				...data,
-				...gitInfo,
-				title: data.title || article.match(/<h1>(.*?)<\/h1>/)?.[1],
-				summary: data.summary || article.match(/<p>(.*?)<\/p>/)?.[1],
-				regularPath: getRegularPath(item),
-				cover: data.cover || '/blogs/art_img1.png',
-			}
-		})
-	)
-	posts.sort(_compareDate)
-	return posts
-}
+  const summary = YAML.load(fs.readFileSync(`docs/summary.yml`), 'utf8')
+  const summaryI18n = langPath ? YAML.load(fs.readFileSync(`docs${langPath}/summary-i18n.yml`), 'utf8') : null
 
-function getRegularPath(str) {
-	let temp = str.split('blogs')
-	let res = `/blogs${temp[temp.length - 1]}`
-	return res.replace('.md', '.html')
-}
+  function makeSidebarItem(items, path, level = 0) {
+    if (Array.isArray(items)) {
+      return items.map(item => makeSidebarItem(item, path, level + 1))
+    } else if (typeof items === 'object') {
+      let title = Object.keys(items)[0]
+      let content = Object.values(items)[0]
 
-function _compareDate(obj1, obj2) {
-	return obj1.date < obj2.date ? 1 : -1
-}
+      if (summaryI18n && !summaryI18n[title]) {
+        return {}
+      }
 
-async function getPostMDFilePaths() {
-	let paths = await globby(['docs/blogs/**.md'], {
-		ignore: ['docs/blogs/index.md', 'README.md'],
-	})
-	return paths
-}
+      return {
+        text: summaryI18n?.[title] || title.replace(/-/g, ' '),
+        items: content.map(item => makeSidebarItem(item, `${path}/${title}`, level + 1)),
+        collapsed: true,
+      }
+    } else {
+      try {
+        let link = `${path}/${items}`.toLocaleLowerCase()
+        let file = fs.readFileSync(`docs${langPath}${link}.md`, 'utf-8')
+        return {
+          text: file.split('\n')[0].replace('# ', ''),
+          link,
+        }
+      } catch (error) {
+        return {}
+      }
+    }
+  }
 
-function getGitTimestamp(file) {
-	return new Promise((resolve, reject) => {
-		const child = spawn('git', ['log', '-1', '--pretty="%ci,%ce,%cn"', file])
-		let output = ''
-		child.stdout.on('data', (d) => (output += String(d)))
-		child.on('close', () => {
-			let [modify, email, author] = output.slice(1, -2).split(',')
-			const avatar = email ? `https://www.gravatar.com/avatar/${md5(email)}` : '/blogs/avatar.png'
-			modify = dayjs(modify).format('MMMM DD , YYYY')
-			author = author || 'Greptime'
-			resolve({
-				modify,
-				email,
-				avatar,
-				author,
-			})
-		})
-		child.on('error', reject)
-	})
-}
-
-export async function getPostLength() {
-	// getPostMDFilePath return type is object not array
-	return [...(await getPostMDFilePaths())].length
+  return makeSidebarItem(summary, '')
 }
